Move vcs source subtypes into their own file
